Fix erroneous CSS output after the last Style Dictionary upgrade (#91)
* Upgrade Style Dictionary and Tokens Studio Transforms

The font size rounding errors that we were seeing are fixed in version 1.1.0 of Tokens Studio Transforms. However, upgrading to this version also requires upgrading to a new major version of Style Dictionary at the same time. The upgrade isn't trivial as there are quite a few breaking changes, but it's good to get it over with.

* Work around some math expressions lacking calc()