Remove color/text/placeholder token
As per https://github.com/element-hq/element-web/issues/28426#issuecomment-2472911879, the color/text/placeholder token is being removed from the design system as it does not generally result in accessible contrast levels. It should be replaced with color/text/secondary.

Note that this update to the token data was performed manually as Aaron and I need to take a look at what's happened with the Tokens Studio setup (and I'm incidentally interested in switching our export format to the new W3C DTCG standard).